backend/requirements.txt
fastapi
uvicorn
python-multipart
pypdf
python-dotenv
langchain
langchain-openai
openai
pydantic